# openapi.yaml
openapi: 3.1.0
info:
  title: Lead intake pipeline
  version: 1.0.0
  description: >
    Accepts a free-text enquiry from any channel and returns immediately. The
    lead is normalized and delivered asynchronously; poll GET /v1/leads/{id} to
    watch it move through the pipeline.

    This document is the contract, not a description of it — the request schema
    here and the zod schema in src/schema.ts describe the same body, and a drift
    check in CI fails the build when they disagree.

servers:
  - url: http://localhost:3210

paths:
  /v1/leads:
    post:
      summary: Submit an enquiry
      description: >
        Idempotent. Supply `Idempotency-Key` (or `idempotency_key`) when the
        caller knows the delivery id; otherwise the server derives one from
        channel + contact + text, which absorbs a duplicated webhook.
      parameters:
        - in: header
          name: Idempotency-Key
          required: false
          schema: { type: string, minLength: 8, maxLength: 200 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LeadInput" }
      responses:
        "202":
          description: Accepted and queued
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LeadAck" }
        "200":
          description: Duplicate — the existing lead is returned unchanged
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LeadAck" }
        "400":
          description: >
            Invalid request body, or an `Idempotency-Key` header that does not
            meet the same length rules as the field.
    get:
      summary: List the 20 most recent leads
      responses:
        "200": { description: OK }

  /v1/leads/{id}:
    get:
      summary: Fetch a lead with its pipeline steps
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Lead" }
        "404": { description: Not found }

  /v1/review:
    get:
      summary: Leads held for a person, with what the pipeline could not answer
      responses:
        "200": { description: OK }

  /v1/review/agreement:
    get:
      summary: How often a person left the extractor's answer alone, per field
      description: >
        Also reports which fields have earned their way out of the queue under
        the configured threshold, and are therefore no longer asked about.
      responses:
        "200": { description: OK }

  /v1/review/{id}:
    get:
      summary: One held lead — the raw text, the flags, and any prior decisions
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200": { description: OK }
        "404": { description: Not found }
    post:
      summary: Record decisions and send the lead back through the pipeline
      description: >
        The merged result is validated against ExtractedLead before anything is
        stored, so a correction cannot put a value into the CRM that the model
        itself would not have been allowed to return. Fields with no empty value
        in that schema cannot be rejected — correct them instead.
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DecisionInput" }
      responses:
        "202": { description: Recorded and re-queued }
        "400": { description: Malformed, or a value the CRM contract refuses }
        "404": { description: Not found }
        "409": { description: This lead is not awaiting review }

  /v1/budget:
    get:
      summary: Rolling 24h model spend against the configured ceiling
      responses:
        "200": { description: OK }

  /health:
    get:
      summary: Liveness plus a database round trip
      responses:
        "200": { description: OK }
        "503": { description: Database unreachable }

components:
  schemas:
    LeadInput:
      type: object
      required: [channel, text]
      additionalProperties: false
      properties:
        channel:
          type: string
          enum: [web_form, whatsapp, email, instagram]
        text:
          type: string
          minLength: 1
          maxLength: 8000
          description: Whatever the customer actually wrote, untouched.
        contact:
          type: string
          maxLength: 200
        idempotency_key:
          type: string
          minLength: 8
          maxLength: 200

    LeadAck:
      type: object
      required: [id, status, duplicate]
      properties:
        id: { type: integer }
        status: { type: string }
        duplicate: { type: boolean }

    ExtractedLead:
      type: object
      required:
        [
          customer_name,
          contact,
          service,
          vehicle_count,
          vehicle_types,
          requested_date,
          urgency,
          language,
          notes,
        ]
      additionalProperties: false
      properties:
        customer_name: { type: [string, "null"] }
        contact: { type: [string, "null"] }
        service:
          type: string
          enum: [wash, detailing, repair, inspection, subscription, other]
        vehicle_count: { type: integer, minimum: 1, maximum: 50 }
        vehicle_types:
          type: array
          items: { type: string }
        requested_date:
          type: [string, "null"]
          pattern: "^\\d{4}-\\d{2}-\\d{2}$"
        urgency:
          type: string
          enum: [asap, today, this_week, flexible]
        language:
          type: string
          enum: [es, en, other]
        notes: { type: string }

    Lead:
      type: object
      properties:
        id: { type: integer }
        channel: { type: string }
        raw_text: { type: string }
        status:
          type: string
          enum: [queued, processing, needs_review, done, failed]
        extracted:
          oneOf:
            - { $ref: "#/components/schemas/ExtractedLead" }
            - { type: "null" }
        extraction_source:
          type: [string, "null"]
          description: >
            `model` when the LLM produced the fields, `heuristic` when the
            deterministic parser did (refusal, truncation, or budget ceiling).
        review_flags:
          description: >
            What the last run could not answer on its own. Null once nothing is
            outstanding.
          oneOf:
            - { type: "null" }
            - type: array
              items: { $ref: "#/components/schemas/ReviewFlag" }
        steps:
          type: array
          items:
            type: object
            properties:
              name: { type: string }
              ok: { type: boolean }
              detail: { type: string }
              ms: { type: integer }

    ReviewFlag:
      type: object
      required: [field, reason, note]
      properties:
        field:
          type: string
          enum: [customer_name, contact, service, vehicle_count, requested_date, urgency]
        reason:
          type: string
          enum: [unreachable, unclassified, impossible_date, disagreement]
        value:
          description: What the pipeline proposes.
        alternative:
          description: What the deterministic parser read, when the two disagreed.
        note:
          type: string
          description: One line, shown to whoever works the queue.

    DecisionInput:
      type: object
      required: [decisions]
      additionalProperties: false
      properties:
        decisions:
          type: array
          minItems: 1
          items:
            type: object
            required: [field, verdict]
            properties:
              field:
                type: string
                enum:
                  [
                    customer_name,
                    contact,
                    service,
                    vehicle_count,
                    requested_date,
                    urgency,
                  ]
              verdict:
                type: string
                enum: [accepted, corrected, rejected]
              value:
                description: >
                  What stands. Omitted for `rejected`, which is a tombstone: the
                  field is emptied and stays empty however many times a later
                  extraction finds the value again.